Rename old cantaloupe module
